MCP/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
gitdb==4.0.12
GitPython==3.1.44
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.7.1
mdurl==0.1.2
narwhals==1.38.1
numpy==2.2.5
openai==1.77.0
openai-agents==0.0.14
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==6.30.2
psutil==7.0.0
psycopg2-binary==2.9.10
pyarrow==20.0.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyMuPDF==1.25.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.3.4
starlette==0.46.2
streamlit==1.45.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.3
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uv==0.7.3
uvicorn==0.34.2
watchdog==6.0.0